Cleanup
A little bit of cleanup, mostly unused imports since they will be automatically added if needed

Commented code is code that wasn't used but not sure if it is needed.

Added destroy to images so they can be garbage collected since you read the image straight from the file anyway (This can be optimised?) but it works.
Moved it to a separate function since you called that block of text twice.

And my formatter appears to have had it's say in some things, if you mind I can unstage the formatting for next time

Moved the background rendering